fix(billing): address Codacy findings — static-content config, Pinia setup, dedup ledger call
- Rename billing.development.config.js → billing.static-content.js (canonical source);
  keep old file as backward-compat re-export shim for downstream projects
- Move useAuthStore/useBillingStore init from computed into setup() in app.vue
  (fixes Pinia store initialised on every computed evaluation — HIGH)
- Align meterAlert texts with i18n key values in billing.alerts.threshold80/100
- Add i18n key comments to hardcoded tab labels in billing.pricing.view.vue
- Remove redundant fetchExtrasLedger() in mounted() — already handled by
  immediate watcher in setup() (billing.subscriptions.component.vue)
